RTA1 sets the new Platinum Standard in floating architecture
        network traffic performance leaps with RTA1 SMP

        best recorded timings for RTA1 TCP traffic
        up from 25'000 responses per second to 32'000

        up to 8 million remote sessions on 8 million sustained TCP connections

        RTA1 emulated machine runs millions of TCP connections

        RTA1 emulator on Linux 32-bit drives 500'000 TCP connections very fast
        RTA1 emulator on Linux 64-bit drives 4'000'000 TCP connections very fast
        RTA1 emulator on macOS 32-bit drives 500'000 TCP connections very fast
        RTA1 emulator on macOS 64-bit drives 8'000'000 TCP connections

+++ rta/tgen.smp/tipt64.c
@@ -50,7 +50,7 @@ static step_second32		 step_second;
 
 static unsigned int		 clockr[2];
 
-static int			 interval_seconds_mask;
+static long			 interval_seconds_mask;
 static int			 platform_interval = PLATFORM_INTERVAL;
 
 /********************************************************
